Fix postgresql_enable logic
Right now, this if will raise an exception if both postgresql_enable and
mysql_enable are false, despite this being a valid condition. Removing
the "or !mysql_enable" from the conditional allows it to work properly,
and there is already a check to prevent mysql_enable and
postgresql_enable from both being true.